refactor: drop dead code and the data/character re-export
2つ整理した。

EmokloreActor#adjustResource はどこからも呼ばれていなかった。ダメージ適用は
applyDamage、共鳴値は raiseResonance、それ以外はシートからの直接編集で足りている。
applyDamage の JSDoc がこれを引き合いに出していたので、そちらも書き直した。

data/character.ts が SkillRef / resolveSkillRef / SkillRollContext / CustomSkillEntry を
character-like.ts から再輸出していた。実体の場所と import 元が食い違い、同じ型の
import 元が2種類に割れていたので、再輸出をやめて character-like.ts に揃えた。
